+++ src/Helpers/functions.php
@@ -1,9 +1,4 @@
 <?php
-/**
- * User: alec
- * Date: 12.10.18
- * Time: 13:21
- */
 
 namespace AlecRabbit;