+++ CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.5.0] - 2019-02-22 
+### Added
+Constants representing time intervals
+```php
+define('AlecRabbit\Helpers\Constants\I_01MIN', 60);
+define('AlecRabbit\Helpers\Constants\I_03MIN', 180);
+define('AlecRabbit\Helpers\Constants\I_05MIN', 300);
+define('AlecRabbit\Helpers\Constants\I_15MIN', 900);
+define('AlecRabbit\Helpers\Constants\I_30MIN', 1800);
+define('AlecRabbit\Helpers\Constants\I_45MIN', 2700);
+define('AlecRabbit\Helpers\Constants\I_01HOUR', 3600);
+define('AlecRabbit\Helpers\Constants\I_02HOUR', 7200);
+define('AlecRabbit\Helpers\Constants\I_03HOUR', 10800);
+define('AlecRabbit\Helpers\Constants\I_04HOUR', 14400);
+define('AlecRabbit\Helpers\Constants\I_01DAY', 86400);
+
+```
 ### Removed
 - function `bc_bounds()`