+++ CHANGELOG.md
@@ -7,7 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.0] - 2019-0.-.. - upcoming release
+## [0.5.0] - 2019-02-22 
+### Added
+Constants representing time intervals
+```php
+define('AlecRabbit\Helpers\Constants\I_01MIN', 60);
+define('AlecRabbit\Helpers\Constants\I_03MIN', 180);
+define('AlecRabbit\Helpers\Constants\I_05MIN', 300);
+define('AlecRabbit\Helpers\Constants\I_15MIN', 900);
+define('AlecRabbit\Helpers\Constants\I_30MIN', 1800);
+define('AlecRabbit\Helpers\Constants\I_45MIN', 2700);
+define('AlecRabbit\Helpers\Constants\I_01HOUR', 3600);
+define('AlecRabbit\Helpers\Constants\I_02HOUR', 7200);
+define('AlecRabbit\Helpers\Constants\I_03HOUR', 10800);
+define('AlecRabbit\Helpers\Constants\I_04HOUR', 14400);
+define('AlecRabbit\Helpers\Constants\I_01DAY', 86400);
+```
+### Removed
+- function `bc_bounds()` 
+
+## [0.4.2] - 2019-02-22 
 ### Added
 - function `array_unset_last()` 
 
@@ -37,8 +56,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.0.16 - 2018-11-29
 
 
-[Unreleased]: https://github.com/alecrabbit/php-helpers/compare/0.3.1...HEAD
-[0.4.0]: https://github.com/alecrabbit/php-helpers/compare/0.3.1...0.4.0
+[Unreleased]: https://github.com/alecrabbit/php-helpers/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/alecrabbit/php-helpers/compare/0.4.2...0.5.0
+[0.4.2]: https://github.com/alecrabbit/php-helpers/compare/0.3.1...0.4.2
 [0.3.1]: https://github.com/alecrabbit/php-helpers/compare/0.2.6...0.3.1
 [0.2.6]: https://github.com/alecrabbit/php-helpers/compare/0.1.4...0.2.6
 [0.1.4]: https://github.com/alecrabbit/php-helpers/compare/0.0.16...0.1.4